Reconcile TableIOServiceTest and TableIOPluginTest
